+++ .github/CHANGELOG.md
@@ -12,6 +12,8 @@
 
 * Adds the function `hafnian_banded` to calculate the hafnian of a banded matrix [#246](https://github.com/XanaduAI/thewalrus/pull/246)
 
+* Adds the functions `hermite_multidimensional_numba` and `grad_hermite_multidimensional_numba` to calculate renormalized multidimensional Hermite polynomials and its gradients using numba [#251](https://github.com/XanaduAI/thewalrus/pull/251)
+
 * Adds the functions `mzgate` and `grad_mzgate` to calculate the Fock representation of the Mach-Zehnder gate and its gradients [#257](https://github.com/XanaduAI/thewalrus/pull/257)
 
 * Adds the ability to calculate n-body photon number distributions using the function `n_body_marginals` [#253](https://github.com/XanaduAI/thewalrus/pull/253)
@@ -40,7 +42,7 @@
 
 This release contains contributions from (in alphabetical order):
 
-Jake Bulmer, Timjan Kalajdzievski, Filippo Miatto, Nicolas Quesada
+Jake Bulmer, Timjan Kalajdzievski, Filippo Miatto, Nicolas Quesada, Yuan Yao
 
 ---